+++ src/Scorpian/Network/RPC/SenderInfo.cs
@@ -1,11 +1,13 @@
+using System;
+
 namespace Scorpian.Network.RPC;
 
 public readonly struct SenderInfo
 {
-    internal SenderInfo(ushort senderId)
+    internal SenderInfo(uint senderId)
     {
         Id = senderId;
     }
     
-    public ushort Id { get; }
+    public uint Id { get; }
 }
