+++ src/Scorpian/SceneManagement/PacketHandlers/SwitchSceneHandler.cs
@@ -9,7 +9,7 @@ public class SwitchSceneHandler : IPacketHandler
     public Type Type => typeof(SwitchScenePacket);
     public NetworkMode Receiver => NetworkMode.Client;
 
-    public void Process(ISyncPacket syncPacket, NetworkedScene networkedScene, ushort senderId, NetworkedSceneManager sceneManager)
+    public void Process(ISyncPacket syncPacket, NetworkedScene networkedScene, uint senderId, NetworkedSceneManager sceneManager)
     {
         var packet = (SwitchScenePacket) syncPacket;
